change the get chip buying and cash out rate showing

+++ app/src/main/java/com/ordering/bet_rate_change/model/ChipTransaction.java
@@ -4,24 +4,24 @@ import com.google.gson.annotations.SerializedName;
 
 public class ChipTransaction {
     @SerializedName("chip_Buying")
-    private int chip_Buying;
+    private double chip_Buying;
 
     @SerializedName("cash_Out")
-    private int cash_Out;
+    private double cash_Out;
 
-    public int getChip_Buying() {
+    public double getChip_Buying() {
         return chip_Buying;
     }
 
-    public void setChip_Buying(int chip_Buying) {
+    public void setChip_Buying(double chip_Buying) {
         this.chip_Buying = chip_Buying;
     }
 
-    public int getCash_Out() {
+    public double getCash_Out() {
         return cash_Out;
     }
 
-    public void setCash_Out(int cash_Out) {
+    public void setCash_Out(double cash_Out) {
         this.cash_Out = cash_Out;
     }
 }
